+++ segmentation_models_pytorch/losses/dice.py
@@ -92,7 +92,7 @@ class DiceLoss(_Loss):
                 y_true = F.one_hot((y_true * mask).to(torch.long), num_classes)  # N,H*W -> N,H*W, C
                 y_true = y_true.permute(0, 2, 1) * mask.unsqueeze(1)  # H, C, H*W
             else:
-                y_true = F.one_hot(y_true, num_classes)  # N,H*W -> N,H*W, C
+                y_true = F.one_hot(y_true.to(torch.long), num_classes)  # N,H*W -> N,H*W, C
                 y_true = y_true.permute(0, 2, 1)  # H, C, H*W
 
         if self.mode == MULTILABEL_MODE:
